+++ sdk/communication/azure-communication-phonenumbers/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release History
 
+## 1.3.0b1 (2025-05-21)
+
+### Features Added
+- Adds support for the Browse Available Phone Numbers and Reservations APIs.
+  - This adds an alternate way to search and purchase phone numbers that allows customers to select which phone numbers they want to reserve and purchase.
+- Adds support for automated purchases of phone numbers from countries requiring a Do Not Resell agreement.
+  - For more information, refer to: https://learn.microsoft.com/azure/communication-services/concepts/numbers/sub-eligibility-number-capability
+- API version `2025-04-01` is the default.
+
 ## 1.2.0 (2025-02-11)
 
 ### Features Added